Improve NAT traversal
Extends the server with a UDP socket to determine peer audio addresses.
This also necessitated an extension to the peer to server connection
protocol with an extra step to determine the correct audio address as
seen from the outside. In the process, the server crate was rewritten in
asynchronous Rust code using Tokio.